refactor: move constants and utils to separate files
