+++ __mocks__/data.ts
@@ -5,4 +5,4 @@ export const MOCK_CARS_ARRAY = [
 
 export const MOCK_SINGLE_CAR = { color: 'black', id: 1, name: 'Ferrari' };
 
-export const MOCK_ENGINE_STATE = { velocity: 64, distance: 100 };
+export const MOCK_ENGINE_STATE = { distance: 100, velocity: 64 };
