--- __mocks__/mock-test-functions.ts
@@ -1,7 +1,0 @@
-export function testTypeGuard<T>(result: boolean) {
-  return function (v: unknown): v is T {
-    void v;
-
-    return result;
-  };
-}
